:building_construction: common code
