+++ packages/client/src/lib/polyfill.ts
@@ -1,0 +1,3 @@
+import { consolehook } from 'vsit'
+
+globalThis.consolehook = consolehook
