+++ packages/client/src/pages/index.tsx
@@ -1,10 +1,10 @@
 /* eslint-disable react/jsx-key */
+import '@/lib/polyfill'
+
 import clsx from 'clsx'
 import Hook from 'console-feed/lib/Hook'
-import { Decode } from 'console-feed/lib/Transform'
 import { createSignal } from 'solid-js'
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { consolehook, MESSAGE_EVENT_TYPE } from 'vsit'
+// @ts-expect-error -- TODO: Fix this type
 import { WBE_API_PATH } from 'vsit-shared/constants'
 
 import { VsitCmdk } from '@/components/cmdk'
@@ -12,71 +12,14 @@ import { CodeMirror } from '@/components/console-feed/codemirror'
 import { fromConsoleToString, removeRemainKeys } from '@/components/console-feed/from-code-to-string'
 import { Resizer } from '@/components/resizeable/resizer'
 import { VsitProvider } from '@/components/vsit-context'
+import { useWS } from '@/hooks/use-ws'
 import { apis } from '@/lib/apis'
-import { VIRUTAL_WEB_ID } from '@/lib/constants'
+import { InitialCode, VIRUTAL_WEB_ID } from '@/lib/constants'
 import { format } from '@/lib/prettier'
 import { withQuery } from '@/lib/utils'
 
-import type { Setter } from 'solid-js'
+import type { Message } from '@/hooks/use-ws'
 
-interface UseWSProps {
-  onMessageUpdate: Setter<Message[]>
-}
-
-const useWS = (props: UseWSProps) => {
-  let socket
-  const importMetaUrl = new URL(import.meta.url)
-  // use server configuration, then fallback to inference
-  const socketProtocol = null || (importMetaUrl.protocol === 'https:' ? 'wss' : 'ws')
-  const hmrPort = null
-  const socketHost = `${null || importMetaUrl.hostname}:${hmrPort || importMetaUrl.port}${'/'}`
-  try {
-    let fallback
-    // only use fallback when port is inferred to prevent confusion
-    // eslint-disable-next-line unused-imports/no-unused-vars
-    socket = setupWebSocket(socketProtocol, socketHost, fallback)
-  } catch (error) {
-    console.error(error)
-  }
-  function setupWebSocket(protocol: string, hostAndPath: string, onCloseWithoutOpen?: () => void) {
-    const socket = new WebSocket(`${protocol}://${hostAndPath}`, 'vite-hmr')
-    let isOpened = false
-    socket.addEventListener('open', () => {
-      console.log('[vit] websocket opened')
-      isOpened = true
-    }, { once: true })
-    // Listen for messages
-    socket.addEventListener('message', async ({ data }) => {
-      const result = JSON.parse(data)
-      if (result.event === MESSAGE_EVENT_TYPE) {
-        const encodeMessage = Decode(Array.isArray(result.data) ? result.data[0] : result.data)
-        props.onMessageUpdate?.([encodeMessage])
-      }
-    })
-    // ping server
-    socket.addEventListener('close', async ({ wasClean }) => {
-      if (wasClean) {
-        return
-      }
-      if (!isOpened && onCloseWithoutOpen) {
-        onCloseWithoutOpen()
-        return
-      }
-      console.log('server connection lost. polling for restart...')
-      // await waitForSuccessfulPing(protocol, hostAndPath);
-      location.reload()
-    })
-    return socket
-  }
-}
-
-globalThis.consolehook = consolehook
-type Message = ReturnType<typeof Decode>
-const InitialCode = `import { uniq } from "esm.sh:lodash-es@4.17.21"
-const a = uniq([1, 2, 3, 3])
-const b: number = 1
-console.log(a, b, uniq)
-`
 const Home = () => {
   const [type, setType] = createSignal<'node' | 'web'>('web')
   const [code, setCode] = createSignal(InitialCode)
